TOOL: renames mappers files to url-mappers
